refactor: config.json imports
