+++ lib/laserAccess.js
@@ -14,7 +14,8 @@ const Bluebird = require('bluebird')
 const CryptoJS = require('crypto-js')
 const debug = require('debug')('laser:control')
 
-const config = require('./config')
+const config = require('../config.json')
+
 const Led = require('./led').Led
 const { maintenanceStatus } = require('./mqtt')
