+++ lib/nomos.js
@@ -5,7 +5,7 @@ const debug = require('debug')('laser:nomos')
 const agent = require('superagent-promise')
 const _ = require('underscore')
 
-const config = require('./config')
+const config = require('../config.json')
 
 module.exports.checkAuth = function (service, id) {
   return agent('POST', config.mmpUrl)
